ARM: 7583/1: decompressor: Enable unaligned memory access for v6 and above
Modern GCC can generate code which makes use of the CPU's native unaligned memory access capabilities.  This is useful for the C decompressor implementations used for unpacking compressed kernels.

This patch disables alignment faults and enables the v6 unaligned access model on CPUs which support these features (i.e., v6 and later), allowing full unaligned access support for C code in the decompressor.

The decompressor C code must not be built to assume that unaligned access works if support for v5 or older platforms is included in the kernel.

For correct code generation, C decompressor code must always use the get_unaligned and put_unaligned accessors when dealing with unaligned pointers, regardless of this patch.