sched: Optimize build_sched_domains() for saving first SD node for a cpu
We are saving first scheduling domain for a cpu in build_sched_domains() by iterating over the nested sd->child list. We don't actually need to do it this way.

*per_cpu_ptr(d.sd, i) is guaranteed to be NULL in the beginning as we have called __visit_domain_allocation_hell() which does a memset to zero for struct s_data.

tl will be equal to sched_domain_topology for the first iteration and so we can set *per_cpu_ptr(d.sd, i) based on that.

So, save pointer to first SD while running the iteration loop over tl's.